// src/app/shared/model/industry-type-category-obj.model.ts
export class IndustryTypeCategoryObj {
  RefIndustryTypeCategoryId: number;
  RefIndustryTypeCategoryCode: string;
  RefIndustryTypeCategoryName: string;
  RefEconomicSectorId: number;
  RegRptCode : string;
  IsActive: any;
  RowVersion: any;
  constructor() { this.RefIndustryTypeCategoryId = 0, this.RowVersion = "" }
}